LayoutTests/platform/ios-simulator/fast/events/click-count-expected.txt

Click inside this div
Clicking 1 times
[Mouse Down] Button: 0 Click Count: 1
[Mouse Up] Button: 0 Click Count: 1
[Mouse Click] Button: 0 Click Count: 1
Clicking 2 times
[Mouse Down] Button: 0 Click Count: 1
[Mouse Up] Button: 0 Click Count: 1
[Mouse Click] Button: 0 Click Count: 1
[Mouse Down] Button: 0 Click Count: 1
[Mouse Up] Button: 0 Click Count: 1
[Mouse Click] Button: 0 Click Count: 1
Clicking 3 times
[Mouse Down] Button: 0 Click Count: 1
[Mouse Up] Button: 0 Click Count: 1
[Mouse Click] Button: 0 Click Count: 1
[Mouse Down] Button: 0 Click Count: 1
[Mouse Up] Button: 0 Click Count: 1
[Mouse Click] Button: 0 Click Count: 1
[Mouse Down] Button: 0 Click Count: 1
[Mouse Up] Button: 0 Click Count: 1
[Mouse Click] Button: 0 Click Count: 1
Clicking 4 times
[Mouse Down] Button: 0 Click Count: 1
[Mouse Up] Button: 0 Click Count: 1
[Mouse Click] Button: 0 Click Count: 1
[Mouse Down] Button: 0 Click Count: 1
[Mouse Up] Button: 0 Click Count: 1
[Mouse Click] Button: 0 Click Count: 1
[Mouse Down] Button: 0 Click Count: 1
[Mouse Up] Button: 0 Click Count: 1
[Mouse Click] Button: 0 Click Count: 1
[Mouse Down] Button: 0 Click Count: 1
[Mouse Up] Button: 0 Click Count: 1
[Mouse Click] Button: 0 Click Count: 1
Clicking 5 times
[Mouse Down] Button: 0 Click Count: 1
[Mouse Up] Button: 0 Click Count: 1
[Mouse Click] Button: 0 Click Count: 1
[Mouse Down] Button: 0 Click Count: 1
[Mouse Up] Button: 0 Click Count: 1
[Mouse Click] Button: 0 Click Count: 1
[Mouse Down] Button: 0 Click Count: 1
[Mouse Up] Button: 0 Click Count: 1
[Mouse Click] Button: 0 Click Count: 1
[Mouse Down] Button: 0 Click Count: 1
[Mouse Up] Button: 0 Click Count: 1
[Mouse Click] Button: 0 Click Count: 1
[Mouse Down] Button: 0 Click Count: 1
[Mouse Up] Button: 0 Click Count: 1
[Mouse Click] Button: 0 Click Count: 1
